+++ frontend/tests/.server/utils/route-utils.test.ts
@@ -1,0 +1,24 @@
+import { describe, expect, it } from 'vitest';
+
+import { i18nRedirect } from '~/.server/utils/route-utils';
+import { ErrorCodes } from '~/errors/error-codes';
+
+describe('route-utils', () => {
+  describe('i18nRedirect', () => {
+    it('should throw an error if no language is found', () => {
+      expect(() => i18nRedirect('routes/public/index.tsx', '')).toThrow(
+        expect.objectContaining({
+          msg: 'No language found in request',
+          errorCode: ErrorCodes.NO_LANGUAGE_FOUND,
+        }),
+      );
+    });
+
+    it('should redirect to the correct path if language is found', () => {
+      const result = i18nRedirect('routes/public/index.tsx', '/en/');
+
+      expect(result.status).toBe(302);
+      expect(result.headers.get('location')).toBe('/en/public');
+    });
+  });
+});
